Use external xmonad-jc-extra in xmonad config
Most of the nix was generated by AI (mostly claude, some chatgpt)

I moved the conditional picture-in-picture layout into a new package
(xmonad-jc-extra) and then use that in my main config. It is included
in both the nixos config and in the devshell I use for working on my
xmonad config.